requirements.txt
gradio
numpy
torch
torchvision
matplotlib
pillow